pyproject.toml
[project]
name = "release-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.123.8",
    "langchain>=1.1.3",
    "langchain-deepseek>=1.0.1",
    "langchain-groq>=1.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
